# openapi.yaml
openapi: 3.1.0
info:
  title: BrainBook API
  description: |
    BrainBook is a social networking platform API that provides functionality for user management, 
    profiles, posts, comments, messaging, and groups. All endpoints follow RESTful conventions 
    and return JSON responses.
    
    ## Route Prefixes
    - `/v1/` - Public routes
    - `/protected/v1/` - Protected routes (requires authenticated user)
    
    ## Authentication
    Authentication is handled via session tokens. After successful login, a session token is 
    created and should be included in subsequent requests to protected endpoints.
  version: 1.0.0
  contact:
    name: BrainBook API Support
  license:
    name: MIT
servers:
  - url: https://api.brainbook.example.com
    description: Production server
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /protected/v1/users/{user_id}/follow:
    post:
      tags:
        - Follow
      summary: Send follow request
      description: |
        Sends a follow request to another user. For public profiles, status is automatically accepted. For private profiles, status is pending.
      operationId: sendFollowRequest
      security:
        - sessionAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Follow request sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '400':
          description: Invalid user ID
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /protected/v1/users/{user_id}/unfollow:
    post:
      tags:
        - Follow
      summary: Unfollow user
      description: Removes an accepted follow relationship between the authenticated user and the target user.
      operationId: unfollowUser
      security:
        - sessionAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Unfollowed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: unfollowed
        '400':
          description: Invalid user ID
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /protected/v1/follow-requests/{request_id}:
    post:
      tags:
        - Follow
      summary: Respond to follow request
      description: Accept or decline a follow request.
      operationId: respondFollowRequest
      security:
        - sessionAuth: []
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [accept, decline]
              required:
                - action
      responses:
        '200':
          description: Follow request response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '400':
          description: Invalid request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /protected/v1/groups/{id}/events:
    get:
      tags:
        - Groups
      summary: Fetch group events
      description: Lists scheduled events for the group. Requires membership.
      operationId: fetchGroupEvents
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/GroupEvent'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Groups
      summary: Create group event
      description: Creates a new event in a group
      operationId: createGroupEvent
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupEventRequest'
      responses:
        '201':
          description: Group event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupEvent'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /protected/v1/groups/{id}/send:
    post:
      tags:
        - Groups
      summary: Send group invite
      description: Allows the group owner to invite another user directly.
      operationId: sendGroupInvite
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - target_user_id
              properties:
                target_user_id:
                  type: integer
                  format: int64
                  description: ID of the user being invited
      responses:
        '200':
          description: Invite already processed or user already a member
        '201':
          description: Invite created successfully
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':

          /protected/v1/groups/{id}/events/{event_id}/rsvp:
            post:
              tags:
                - Groups
              summary: RSVP to group event
              description: Allows a group member to set their RSVP status (going/not_going) for an event.
              operationId: rsvpGroupEvent
              security:
                - sessionAuth: []
              parameters:
                - name: id
                  in: path
                  required: true
                  description: Group ID
                  schema:
                    type: integer
                - name: event_id
                  in: path
                  required: true
                  description: Event ID
                  schema:
                    type: integer
              requestBody:
                required: true
                content:
                  application/json:
                    schema:
                      type: object
                      properties:
                        response:
                          type: string
                          enum: [going, not_going]
                      required:
                        - response
              responses:
                '200':
                  description: RSVP recorded
                '401':
                  $ref: '#/components/responses/UnauthorizedError'
                '404':
                  description: Group or event not found
                  content:
                    application/json:
                      schema:
                        $ref: '#/components/schemas/ErrorResponse'
                '500':
                  $ref: '#/components/responses/InternalServerError'
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Requester is not the group owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /protected/v1/groups/{group_id}/requests/{request_id}:
    post:
      tags:
        - Groups
      summary: Respond to group join/invite request
      description: Allows the request target (group owner or invited user) to accept or decline a group join request or invitation.
      operationId: respondGroupRequest
      security:
        - sessionAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: integer
        - name: request_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [accept, decline]
              required:
                - action
      responses:
        '200':
          description: Request processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: accepted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Session token obtained after successful login
  schemas:
    ErrorResponse:
      type: object
      properties:
        Error:
          type: string
          description: Error message
      required:
        - Error
...
    GroupEvent:
      type: object
      properties:
        id:
          type: integer
        group_id:
          type: integer
        user_id:
          type: integer
          description: Creator user ID
        title:
          type: string
        description:
          type: string
        time:
          type: string
          format: date-time
        interested:
          type: integer
          description: Number of members marked as going
        not_interested:
          type: integer
          description: Number of members marked as not going
      required:
        - id
        - group_id
        - user_id
        - title
        - time
    CreateGroupEventRequest:
      type: object
      properties:
        title:
          type: string
          description: Event title
        description:
          type: string
          description: Event description
        time:
          type: string
          format: date-time
          description: Date and time of the event
      required:
        - title
        - time
...
  responses:
    UnauthorizedError:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'